add providers to plugin reduce
